Added util struct RgbChannel, util function get_bits(u8)
The RgbChannel is a placeholder to track the current channel of a bit plane as an attribute. When I'll manage to track it in the Block that contains it, it will be removed.

The get_bits function gets a u8 (single byte) and returns the bits of it in left to right order (MSB is index 0, LSB is index 7)